requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.2.7
    # via
    #   -r requirements.in
    #   django
atomicwrites==1.4.0
    # via
    #   -r requirements.in
    #   pytest
attrs==19.3.0
    # via
    #   -r requirements.in
    #   pytest
certifi==2020.4.5.1
    # via
    #   -r requirements.in
    #   requests
chardet==3.0.4
    # via
    #   -r requirements.in
    #   requests
click==7.1.2
    # via
    #   -r requirements.in
    #   pip-tools
colorama==0.4.4
    # via
    #   -r requirements.in
    #   pytest
django-cors-headers==3.7.0
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django==3.0.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.6.0
    # via -r requirements.in
djangorestframework==3.11.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
gunicorn==20.0.4
    # via -r requirements.in
idna==2.9
    # via
    #   -r requirements.in
    #   requests
importlib-metadata==1.6.0
    # via -r requirements.in
more-itertools==8.2.0
    # via
    #   -r requirements.in
    #   pytest
packaging==20.3
    # via
    #   -r requirements.in
    #   pytest
pep517==0.10.0
    # via
    #   -r requirements.in
    #   pip-tools
pip-tools==6.0.1
    # via -r requirements.in
pluggy==0.13.1
    # via
    #   -r requirements.in
    #   pytest
psycopg2-binary==2.8.6
    # via -r requirements.in
py==1.8.1
    # via
    #   -r requirements.in
    #   pytest
pyjwt==2.0.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
pyparsing==2.4.7
    # via
    #   -r requirements.in
    #   packaging
pytest-django==3.9.0
    # via -r requirements.in
pytest==5.4.1
    # via
    #   -r requirements.in
    #   pytest-django
pytz==2019.3
    # via
    #   -r requirements.in
    #   django
requests==2.23.0
    # via -r requirements.in
six==1.14.0
    # via
    #   -r requirements.in
    #   packaging
sqlparse==0.3.1
    # via
    #   -r requirements.in
    #   django
toml==0.10.2
    # via
    #   -r requirements.in
    #   pep517
urllib3==1.25.9
    # via
    #   -r requirements.in
    #   requests
wcwidth==0.1.9
    # via
    #   -r requirements.in
    #   pytest
zipp==3.1.0
    # via
    #   -r requirements.in
    #   importlib-metadata

